requirements.txt
djangorestframework
django
djangorestframework-simplejwt
psycopg2-binary
pillow
drf-yasg
python-decouple
django-cors-headers
redis
celery
django-filter
djangorestframework-pagination
beautifulsoup4
lxml
requests
gunicorn
django-password-validators
redis
setuptools